fix: remove console-warns, rename parser
