+++ src/data-fetchers/bed/bed-data-fetcher.ts
@@ -69,7 +69,6 @@ class BedDataFetcher implements TabularDataFetcher<BedTile> {
     async getTabularData(tileIds: string[]): Promise<BedTile[]> {
         const buf = await (await this.worker).getTabularData(this.uid, tileIds);
         const parsed = JSON.parse(new TextDecoder().decode(buf));
-        console.warn(parsed);
         return parsed;
     }
 }
