+++ src/data-fetchers/bed/bed-parser.test.ts
@@ -1,4 +1,4 @@
-import { BedParser } from './parser';
+import { BedParser } from './bed-parser';
 
 describe('BED parser with default columns', () => {
     const parser = new BedParser();
